+++ src/main/java/stringcalculator/StringAddCalculator.java
@@ -1,4 +1,4 @@
-package string_calculator;
+package stringcalculator;
 
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
