+++ src/test/java/stringcalculator/StringAddCalculatorTest.java
@@ -1,4 +1,4 @@
-package string_calculator;
+package stringcalculator;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
