+++ src/test/java/stringset/SetTest.java
@@ -1,4 +1,4 @@
-package String_Set;
+package stringset;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
